HOM-7 HOM-9 HOM-27 additions

+++ server/data/homes.js
@@ -1,7 +1,8 @@
 import { isValidObjectId } from "mongoose";
 import Home from "../models/home.js";
+import Image from "../models/image.js";
 
-export const getHome = async (userId) => {
+export const getHomes = async (userId) => {
   if (!isValidObjectId(userId))
     throw { status: 400, message: "Error: Invaild User Id" };
 
@@ -9,7 +10,26 @@ export const getHome = async (userId) => {
   return home;
 };
 
-export const deleteHome = async (userId) => {
+export const deleteHomes = async (userId) => {
+  if (!isValidObjectId(userId))
+    throw { status: 400, message: "Error: Invaild User Id" };
   const deletedHomes = await Home.deleteMany({ userId: userId });
   return deletedHomes.deletedCount !== 0;
+};
+
+export const getAllHomeImagesOfUser = async (userId) => {
+  try {
+    const homes = await getHomes(userId);
+    const imagesArray = [];
+    for (let home of homes) {
+      let tempImageArray = await Image.find({
+        imageableId: home._id,
+        imageableType: "Home",
+      });
+      imagesArray.concat(tempImageArray);
+    }
+    return imagesArray;
+  } catch (e) {
+    throw e;
+  }
 };
